+++ lib/mongoid/persistable.rb
@@ -7,6 +7,7 @@ require "mongoid/persistable/incrementable"
 require "mongoid/persistable/logical"
 require "mongoid/persistable/poppable"
 require "mongoid/persistable/pullable"
+require "mongoid/persistable/pushable"
 require "mongoid/persistable/renamable"
 require "mongoid/persistable/savable"
 require "mongoid/persistable/settable"
@@ -30,6 +31,7 @@ module Mongoid
     include Logical
     include Poppable
     include Pullable
+    include Pushable
     include Renamable
     include Savable
     include Settable