+++ lib/mongoid/persistable/atomic.rb
@@ -1,8 +1,6 @@
 # encoding: utf-8
 require "mongoid/persistable/atomic/operation"
 require "mongoid/persistable/atomic/add_to_set"
-require "mongoid/persistable/atomic/push"
-require "mongoid/persistable/atomic/push_all"
 
 module Mongoid
   module Persistable
@@ -31,40 +29,6 @@ module Mongoid
       # @since 2.0.0
       def add_to_set(field, value, options = {})
         AddToSet.new(self, field, value, options).persist
-      end
-
-      # Performs an atomic $push of the provided value on the supplied field. If
-      # the field does not exist it will be initialized as an empty array.
-      #
-      # @example Push a value on the field.
-      #   person.push(:aliases, "Bond")
-      #
-      # @param [ Symbol ] field The name of the field.
-      # @param [ Object ] value The value to push.
-      # @param [ Hash ] options The mongo persistence options.
-      #
-      # @return [ Array<Object> ] The new value of the field.
-      #
-      # @since 2.0.0
-      def push(field, value, options = {})
-        Push.new(self, field, value, options).persist
-      end
-
-      # Performs an atomic $pushAll of the provided value on the supplied field. If
-      # the field does not exist it will be initialized as an empty array.
-      #
-      # @example Push the values onto the field.
-      #   person.push_all(:aliases, [ "Bond", "James" ])
-      #
-      # @param [ Symbol ] field The name of the field.
-      # @param [ Array<Object> ] value The values to push.
-      # @param [ Hash ] options The mongo persistence options.
-      #
-      # @return [ Array<Object> ] The new value of the field.
-      #
-      # @since 2.1.0
-      def push_all(field, value, options = {})
-        PushAll.new(self, field, value, options).persist
       end
     end
   end