+++ spec/mongoid/persistable/atomic_spec.rb
@@ -22,43 +22,5 @@ describe Mongoid::Persistable::Atomic do
         expect(person.reload.array).to eq([ "test", "testy" ])
       end
     end
-
-    describe "#push" do
-
-      let(:person) do
-        Person.create(array: [ "test" ])
-      end
-
-      before do
-        person.push(:array, "testy")
-      end
-
-      it "adds to the aliased field" do
-        expect(person.array).to eq([ "test", "testy" ])
-      end
-
-      it "persists the change" do
-        expect(person.reload.array).to eq([ "test", "testy" ])
-      end
-    end
-
-    describe "#push_all" do
-
-      let(:person) do
-        Person.create(array: [ "test" ])
-      end
-
-      before do
-        person.push_all(:array, [ "testy", "test2" ])
-      end
-
-      it "adds to the aliased field" do
-        expect(person.array).to eq([ "test", "testy", "test2" ])
-      end
-
-      it "persists the change" do
-        expect(person.reload.array).to eq([ "test", "testy", "test2" ])
-      end
-    end
   end
 end